Revert "Compute queryformat during convert of dto"
This reverts commit 473bf70c55038375a24baf3e928032b3da9b3a82.
