+++ internal/cache/cache_manager_test.go
@@ -105,8 +105,7 @@ trackEvents = false
 				"test-flag": {
 					Rules: &[]flag.Rule{
 						{
-							Query:       testconvert.String("key eq \"random-key\""),
-							QueryFormat: "nikunjy",
+							Query: testconvert.String("key eq \"random-key\""),
 							Percentages: &map[string]float64{
 								"false_var": 0,
 								"true_var":  100,
@@ -153,8 +152,7 @@ test-flag:
 				"test-flag": {
 					Rules: &[]flag.Rule{
 						{
-							Query:       testconvert.String("key eq \"random-key\""),
-							QueryFormat: flag.NikunjyQueryFormat,
+							Query: testconvert.String("key eq \"random-key\""),
 							Percentages: &map[string]float64{
 								"false_var": 0,
 								"true_var":  100,
@@ -208,8 +206,7 @@ test-flag:
 				"test-flag": {
 					Rules: &[]flag.Rule{
 						{
-							Query:       testconvert.String("key eq \"random-key\""),
-							QueryFormat: flag.NikunjyQueryFormat,
+							Query: testconvert.String("key eq \"random-key\""),
 							Percentages: &map[string]float64{
 								"false_var": 0,
 								"true_var":  100,
@@ -313,8 +310,7 @@ test-flag:
 					},
 					Rules: &[]flag.Rule{
 						{
-							Query:       testconvert.String("key eq \"random-key\""),
-							QueryFormat: flag.NikunjyQueryFormat,
+							Query: testconvert.String("key eq \"random-key\""),
 							Percentages: &map[string]float64{
 								"false_var": 0,
 								"true_var":  100,
@@ -369,8 +365,7 @@ test-flag2:
 					},
 					Rules: &[]flag.Rule{
 						{
-							Query:       testconvert.String("key eq \"random-key\""),
-							QueryFormat: flag.NikunjyQueryFormat,
+							Query: testconvert.String("key eq \"random-key\""),
 							Percentages: &map[string]float64{
 								"false_var": 0,
 								"true_var":  100,
@@ -389,8 +384,7 @@ test-flag2:
 					},
 					Rules: &[]flag.Rule{
 						{
-							Query:       testconvert.String("key eq \"random-key\""),
-							QueryFormat: flag.NikunjyQueryFormat,
+							Query: testconvert.String("key eq \"random-key\""),
 							Percentages: &map[string]float64{
 								"false_var": 100,
 								"true_var":  0,
