+++ internal/flag/rule.go
@@ -16,6 +16,13 @@ import (
 	"github.com/thomaspoignant/go-feature-flag/internal/utils"
 )
 
+type QueryFormat = string
+
+const (
+	NikunjyQueryFormat   QueryFormat = "nikunjy"
+	JSONLogicQueryFormat QueryFormat = "jsonlogic"
+)
+
 // Rule represents a rule applied by the flag.
 type Rule struct {
 	// Name is the name of the rule, this field is mandatory if you want
@@ -41,8 +48,6 @@ type Rule struct {
 
 	// Disable indicates that this rule is disabled.
 	Disable *bool `json:"disable,omitempty" yaml:"disable,omitempty" toml:"disable,omitempty" jsonschema:"title=disable,description=Indicates that this rule is disabled."` // nolint: lll
-
-	QueryFormat QueryFormat `json:"-" yaml:"-" toml:"-"` // nolint: lll
 }
 
 // Evaluate is checking if the rule applies to for the user.
@@ -356,11 +361,10 @@ func (r *Rule) GetTrimmedQuery() string {
 
 // GetQueryFormat is returning the format used for the query
 func (r *Rule) GetQueryFormat() QueryFormat {
-	if r.QueryFormat != "" {
-		return r.QueryFormat
+	if utils.IsJSONObject(r.GetTrimmedQuery()) {
+		return JSONLogicQueryFormat
 	}
-	r.QueryFormat = GetQueryFormat(*r)
-	return r.QueryFormat
+	return NikunjyQueryFormat
 }
 
 func (r *Rule) GetQuery() string {
